Separated utils/index.js into utils/path.js and utils/translations.js
